Add test pantry has enough ingredients
