+++ lib/cookbook.rb
@@ -24,5 +24,7 @@ class CookBook
     end
   end
 
+  
+
 
 end
