+++ test/cookbook_test.rb
@@ -64,31 +64,7 @@ class CookBookTest < Minitest::Test
     assert_equal recipe2, cookbook.highest_calorie_meal
   end
 
-  def test_enough_ingredients_for?
-    pantry = Pantry.new
-    cookbook = CookBook.new
-    ingredient1 = Ingredient.new({name: "Cheese", unit: "C", calories: 100})
-    ingredient2 = Ingredient.new({name: "Macaroni", unit: "oz", calories: 30})
-    recipe1 = Recipe.new("Mac and Cheese")
-    recipe1.add_ingredient(ingredient1, 2)
-    recipe1.add_ingredient(ingredient2, 8)
-    ingredient3 = Ingredient.new({name: "Ground Beef", unit: "oz", calories: 100})
-    ingredient4 = Ingredient.new({name: "Bun", unit: "g", calories: 75})
-    recipe2 = Recipe.new("Cheese Burger")
-    recipe2.add_ingredient(ingredient1, 2)
-    recipe2.add_ingredient(ingredient3, 4)
-    recipe2.add_ingredient(ingredient4, 1)
-    cookbook.add_recipe(recipe1)
-    cookbook.add_recipe(recipe2)
-    pantry.restock(ingredient1, 5)
-    pantry.restock(ingredient1, 10)
-    assert_equal false, pantry.enough_ingredients_for?(recipe1)
-    pantry.restock(ingredient2, 7)
-    assert_equal false, pantry.enough_ingredients_for?(recipe1)
-    pantry.restock(ingredient2, 1)
-    assert pantry.enough_ingredients_for?(recipe1)
-
-  end
+
